requirements.txt
# GUI框架
PyQt6>=6.4.0
PyQt6-Qt6>=6.4.0

# 数据处理
numpy>=1.21.0
pandas>=1.3.0

# 图像处理
Pillow>=8.3.0
opencv-python>=4.5.0

# 音频处理
pydub>=0.25.0

# 配置文件处理
PyYAML>=6.0
configparser>=5.2.0

# 数据序列化
pydantic>=1.9.0

# 日志处理
loguru>=0.6.0

# 测试框架
pytest>=7.0.0
pytest-qt>=4.0.0

# 文档生成
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# 开发工具
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# 地图文件处理
struct>=3.0.0

# 网络请求
requests>=2.28.0

# 进度条
tqdm>=4.64.0

# 数据可视化
matplotlib>=3.5.0
seaborn>=0.11.0 